区服关系梳理 ， 目前可以 表示 该账号 在哪个区的：
1.登录阶段： 账号登录中心服务器上 记录该账号 去了那个服，
2.进入场景阶段：有了Unit组件可以 获取 UnitIdStruct.GetUnitZone(player.UintId);   这个player.UintId  也是 选择的角色 id， 也是 unitid，也是palyer目前的id，根据 unitid来找出他目前在那个服上。

+++ Server/Model/Generate/Message/OuterMessage.cs
@@ -605,7 +605,7 @@ namespace ET
 		public long AccountId { get; set; }
 
 		[ProtoMember(2)]
-		public long ServerId { get; set; }
+		public int ServerId { get; set; }
 
 		[ProtoMember(3)]
 		public string Name { get; set; }
@@ -642,7 +642,7 @@ namespace ET
 		public string Name { get; set; }
 
 		[ProtoMember(4)]
-		public long ServerId { get; set; }
+		public int ServerId { get; set; }
 
 	}
 
@@ -679,7 +679,7 @@ namespace ET
 		public string Token { get; set; }
 
 		[ProtoMember(4)]
-		public long ServerId { get; set; }
+		public int ServerId { get; set; }
 
 	}
 
@@ -722,7 +722,7 @@ namespace ET
 		public string Name { get; set; }
 
 		[ProtoMember(5)]
-		public long ServerId { get; set; }
+		public int ServerId { get; set; }
 
 	}
 
@@ -756,7 +756,7 @@ namespace ET
 		public string Token { get; set; }
 
 		[ProtoMember(3)]
-		public long ServerId { get; set; }
+		public int ServerId { get; set; }
 
 	}
